+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.Infrastructure/Persistence/Mappers/ActorMapper.cs
@@ -77,7 +77,7 @@ internal static class ActorMapper
 
         var actorNumber = ActorNumber.Create(from.ActorNumber);
         var actorStatus = from.Status;
-        var actorName = new ActorName(from.Name);
+        var actorName = new ActorName(string.IsNullOrWhiteSpace(from.Name) ? "-" : from.Name); // TODO: This check should be removed once we are on new env.
 
         return new Actor(
             new ActorId(from.Id),